move the filter component to molecules
