+++ ui/litellm-dashboard/src/components/all_keys_table.tsx
@@ -7,8 +7,8 @@ import { Button } from "@tremor/react"
 import KeyInfoView from "./templates/key_info_view"
 import { Tooltip } from "antd"
 import { Team, KeyResponse } from "./key_team_helpers/key_list"
-import FilterComponent from "./common_components/filter"
-import { FilterOption } from "./common_components/filter"
+import FilterComponent from "./molecules/filter"
+import { FilterOption } from "./molecules/filter"
 import { keyListCall, Organization, userListCall } from "./networking"
 import { createTeamSearchFunction } from "./key_team_helpers/team_search_fn"
 import { createOrgSearchFunction } from "./key_team_helpers/organization_search_fn"
