+++ ui/litellm-dashboard/src/components/view_logs/index.tsx
@@ -18,8 +18,8 @@ import KeyInfoView from "../templates/key_info_view"
 import { SessionView } from "./SessionView"
 import { VectorStoreViewer } from "./VectorStoreViewer"
 import { GuardrailViewer } from "./GuardrailViewer"
-import FilterComponent from "../common_components/filter"
-import { FilterOption } from "../common_components/filter"
+import FilterComponent from "../molecules/filter"
+import { FilterOption } from "../molecules/filter"
 import { useLogFilterLogic } from "./log_filter_logic"
 import { fetchAllKeyAliases } from "../key_team_helpers/filter_helpers"
 import { Tab, TabGroup, TabList, TabPanels, TabPanel, Text, Switch } from "@tremor/react"
